requirements.txt
# Web 框架核心
Flask==3.1.0
Flask-Cors==5.0.1

# 数据库
neo4j==5.28.1

# 文件处理 - 按需启用
# PyPDF2==2.0.0
# python-docx==0.8.11

# 环境变量
python-dotenv==1.0.0

# HTTP 客户端
requests==2.32.3

# 自定义依赖
kg-gen==0.1.6 